+++ CHANGELOG.md
@@ -6,6 +6,8 @@
     - better log.
     - fix matches order of a query.
     - 12-15% faster.
+- `profile`:
+    - change computation of `score`.
 
 ### v0.6.0 - 2021-08-13